+++ psd-web/app/controllers/investigations_controller.rb
@@ -17,8 +17,10 @@ class InvestigationsController < ApplicationController
       end
       format.xlsx do
         @answer = search_for_investigations
-        @investigations = Investigation.eager_load(:complainant,
-                                                   :source).where(id: @answer.results.map(&:_id))
+        @investigations = Investigation.eager_load(
+          :complainant,
+          :source
+        ).where(id: @answer.results.map(&:_id))
 
         @activity_counts = Activity.group(:investigation_id).count
         @business_counts = InvestigationBusiness.unscoped.group(:investigation_id).count
@@ -82,11 +84,12 @@ private
 
     respond_to do |format|
       if @investigation.update(update_params)
-        format.html {
-          redirect_to investigation_path(@investigation), flash: {
-            success: "#{@investigation.case_type.upcase_first} was successfully updated."
-          }
-        }
+        format.html do
+          redirect_to investigation_path(@investigation),
+                      flash: {
+                        success: "#{@investigation.case_type.upcase_first} was successfully updated."
+                      }
+        end
         format.json { render :show, status: :ok, location: @investigation }
       else
         format.html { render action_name }