+++ psd-web/spec/forms/edit_investigation_collaborator_form_spec.rb
@@ -25,9 +25,9 @@ RSpec.describe EditInvestigationCollaboratorForm, :with_elasticsearch, :with_stu
     }
   end
 
-  let(:form) {
+  let(:form) do
     described_class.new(params)
-  }
+  end
 
   before do
     editor