+++ psd-web/spec/requests/add_collaborator_spec.rb
@@ -65,14 +65,16 @@ RSpec.describe "Adding a collaborator to a case", type: :request, with_stubbed_m
   context "when trying to add a team who is already a collaborator" do
     let(:existing_collaborator_team) { create(:team) }
     let(:investigation) {
-      create(:investigation,
-             owner: user,
-             editions: [
-               create(:edition,
-                      collaborator: existing_collaborator_team,
-                      include_message: false,
-                      added_by_user: user)
-             ])
+      create(
+        :investigation,
+        owner: user,
+        editions: [
+          create(:edition,
+                 collaborator: existing_collaborator_team,
+                 include_message: false,
+                 added_by_user: user)
+        ]
+      )
     }
 
     before do